backend/requirements.txt
urllib3==2.0.7
openai==1.12.0
python-dotenv==1.0.0
Django==4.2.1
aiohttp<=3.8.6,<4.0.0
Brotli==1.1.0
flask==2.3.3
gunicorn==21.2.0
gevent==23.9.0.post1
edge-tts==6.1.8
djangorestframework==3.14.0
django-cors-headers==4.2.0
channels==4.0.0
daphne==4.0.0
schedule==1.2.0
drf-yasg==1.21.7
pymilvus==2.2.12
datetime==5.2
transformers>=4.36.2
huggingface_hub>=0.19.4
sentence-transformers==2.2.2
easygoogletranslate==0.0.4
volcengine==1.0.103
requests==2.31.0
jieba==0.42.1
websockets==11.0.3
APScheduler==3.10.4
zep-python==1.5.0
pypinyin==0.50.0
bilibili-api-python==16.1.1
faiss-cpu==1.7.4
FlagEmbedding==1.1.7
litellm==1.34.6